:sparkles: feat: add PageHeader component (#617)
New PageHeader component for consistent header appearance across pages
