+++ client/flutter/lib/components/page_scaffold.dart
@@ -1,4 +1,4 @@
-import 'package:WHOFlutter/components/back_arrow.dart';
+import 'package:WHOFlutter/components/page_header.dart';
 import 'package:flutter/material.dart';
 
 class PageScaffold extends StatelessWidget {
@@ -31,7 +31,7 @@ class PageScaffold extends StatelessWidget {
                 color: Colors.transparent,
               ),
               backgroundColor: Colors.white,
-              flexibleSpace: FlexibleSpaceBar(background: _buildHeader()),
+              flexibleSpace: PageHeader(context, title: this.title, subtitle: this.subtitle, padding: this.padding, showBackButton: this.showBackButton),
               expandedHeight: 120,
             ),
             ...this.body
@@ -39,43 +39,4 @@ class PageScaffold extends StatelessWidget {
         ));
   }
 
-  SafeArea _buildHeader() {
-    List<Widget> headerItems = [
-      this.showBackButton ? _buildBackArrow() : null,
-      Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Text(this.title,
-              textScaleFactor: 1.8,
-              style: TextStyle(
-                  color: Colors.blueAccent, fontWeight: FontWeight.bold)),
-          SizedBox(height: 4),
-          Text(this.subtitle,
-              textScaleFactor: 1.0,
-              style:
-                  TextStyle(color: Colors.black, fontWeight: FontWeight.bold)),
-        ],
-      ),
-      Image.asset('assets/images/mark.png', width: 75),
-    ];
-    headerItems.removeWhere((element) => element == null);
-    return SafeArea(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 24.0),
-        child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: headerItems),
-      ),
-    );
-  }
-
-  GestureDetector _buildBackArrow() {
-    return GestureDetector(
-      behavior: HitTestBehavior.opaque,
-      onTap: () => Navigator.pop(this.context),
-      child: BackArrow(),
-    );
-  }
 }
