+++ content/credits.yaml
@@ -35,6 +35,7 @@ team:
   - Guido Rosso
   - Luigi Rosso
   - Kristina Plummer
+  - Kieran Uddin
   
 # Add yourself here when you first submit a contribution.
